requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.3.0
PyMySQL==1.1.1
python-dotenv==1.1.0
